maxq works
Key problem was a bug in the completion function update, it was using the best action for the result state to index into the completion table for updating instead of using the action taken.

maxqq was more properly implemented while trying to figure out what was wrong.  And some refactoring.